add server-side admin api authentication
Authenticate protected backend requests without exposing the shared key to browser code.
